Add files via upload
Are you one of these Persons which needs to watch Videos to fall asleep? this is for you. A simple elegant and lightweight Shutdowntimer.
And if you have a Girlfriend which is complaining about the flickering lights, or even if you hate them, you can change the brightness of your screen too!



For now only in GERMAN! i will update to English soon!